Added support for REPEAT command.
This required changing structure of how default delays are set
and retrieved. Now each DuckyCommand instance is given the same
reference to a dictionary holding a DuckyScript's methods for
setting and getting default delay and its attribute that holds
created commands.

Also fixed a bug that caused aliases to behave incorrectly.